Added commons and commons validator Editied Listing Entity and ListingDTO according to the task Added validation utlity Updated ListingRepository with streaming Added logging to loaders Created ReportMaker class

+++ src/main/java/hu/wortex/report/entities/ListingDTO.java
@@ -25,7 +25,7 @@ public class ListingDTO {
     private String locationId;
 
     @JsonProperty("listing_price")
-    private Integer listingPrice;
+    private Long listingPrice;
 
     @JsonProperty("currency")
     private String currency;
